fix(web-channel): race condition from visit event
